+++ src/vm/os/windows/memory.cpp
@@ -571,7 +571,7 @@ uintptr_t WINAPI MemoryManager::CollectMemory(void* arg)
   }
   
   // remove references from allocated pool
-  for(unsigned int i = 0; i < erased_memory.size(); i++) {
+  for(size_t i = 0; i < erased_memory.size(); i++) {
     allocated_memory.erase(erased_memory[i]);
     allocated_int_obj_array.erase(erased_memory[i]);
   }
@@ -591,7 +591,7 @@ uintptr_t WINAPI MemoryManager::CollectMemory(void* arg)
   return 0;
 }
 
-unsigned int WINAPI MemoryManager::CheckStatic(void* arg)
+size_t WINAPI MemoryManager::CheckStatic(void* arg)
 {
   // check static memory
 #ifndef _GC_SERIAL
